unify mocked_ldap
