+++ tests/rdmo/conftest.py
@@ -1,11 +1,5 @@
-from unittest.mock import MagicMock
-from uuid import UUID
-
 import pytest
-from pytest import MonkeyPatch
 
-from mex.common.ldap.connector import LDAPConnector
-from mex.common.ldap.models.person import LDAPPerson
 from mex.common.models import ExtractedPerson
 from mex.common.types import Identifier
 from mex.rdmo.settings import RDMOSettings
@@ -28,24 +22,3 @@ def extracted_person() -> ExtractedPerson:
         hadPrimarySource=Identifier.generate(seed=43),
         identifierInPrimarySource="arnold",
     )
-
-
-@pytest.fixture
-def mocked_ldap(monkeypatch: MonkeyPatch) -> None:
-    """Mock the LDAP connector to return resolved persons and units."""
-    persons = [
-        LDAPPerson(
-            employeeID="42",
-            sn="Resolved",
-            givenName="Roland",
-            displayName="Resolved, Roland",
-            objectGUID=UUID(int=4, version=4),
-            department="PARENT-UNIT",
-        )
-    ]
-    monkeypatch.setattr(
-        LDAPConnector,
-        "__init__",
-        lambda self: setattr(self, "_connection", MagicMock()),
-    )
-    monkeypatch.setattr(LDAPConnector, "get_persons", lambda _, **__: iter(persons))
